[UI/UX] Add Twitter Bootstrap framework
See http://twitter.github.io/bootstrap/ for details on Bootstrap.

Getting this to work on Rails 2.3 is a little more tricky than in
current Rails (3.x) since it doesn't have Sprockets. We workaround this
by using Compass (http://compass-style.org/) which then allows us to use
bootstrap-sass (https://github.com/thomas-mcdonald/bootstrap-sass).

These asset (CSS, Javascript) compilation-only Gem dependencies don't
need to be included in the product as we can precompile the assets
during build time. But we'll need to modify the code to allow this, or
find another solution.